requirements.txt
tika
colorama
python-whois